+++ ChangeLog.md
@@ -1,5 +1,7 @@
 2022-09
 
+- re-enabled google base layers
+  - as in other apps using googlemutant=0.8.0 due to new 0.13 not working right away
 - dispatch DMapDialog in MissionTemplate page
 - set version 0.9.2 toward release
 - initial commit with an overall migration of and improvement over previous version.